+++ plugin_info/install.php
@@ -734,7 +734,7 @@
 
             // Models reset
             // Some Xiaomi devices are now managed thru 'xiaomi' structure in model.
-            $toReload = ['sensor_magnet.aq2', 'sensor_motion.aq2', 'weather', 'motion.ac02', 'switch.n0agl1'];
+            $toReload = ['sensor_magnet.aq2', 'sensor_motion.aq2', 'weather', 'motion.ac02', 'switch.n0agl1', 'sensor_wleak.aq1'];
             foreach ($eqLogics as $eqLogic) {
                 $eqModel = $eqLogic->getConfiguration('ab::eqModel', []);
                 if (!isset($eqModel['id']))
@@ -745,7 +745,7 @@
                 // Check last update from model
                 $updateTime = $eqLogic->getConfiguration('updatetime', "");
                 if ($updateTime != '') {
-                    if (strtotime($updateTime) > strtotime("2023-01-13 16:33:00"))
+                    if (strtotime($updateTime) > strtotime("2023-01-25 16:33:00"))
                         continue;
                 }